requirements.txt
# PyTorch Installation (REQUIRED - install separately before other dependencies)
# PyTorch must be installed first with the appropriate backend for your system.
#
# For CPU-only:
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
# For GPU (CUDA 12.1):
#   pip install torch --index-url https://download.pytorch.org/whl/cu121
#
# See https://pytorch.org/get-started/locally/ for other CUDA versions

# Core Dependencies
transformers>=4.46.0        # Hugging Face ASR pipeline
librosa>=0.10.0             # audio loading and resampling
scipy>=1.10.0               # signal processing for audio preprocessing
pydub>=0.25.1               # MP3 to WAV conversion
python-dotenv>=1.0.0        # environment configuration
tqdm>=4.67.0                # progress bars
pydantic>=2.9.2             # data validation and typing support
openai>=1.0.0               # Azure OpenAI API client
azure-cognitiveservices-speech>=1.35.0  # Azure TTS for audio generation

# Fine-tuning Dependencies
datasets>=2.14.0            # HuggingFace datasets for data loading
peft>=0.7.0                 # LoRA and parameter-efficient fine-tuning
evaluate>=0.4.0             # Metrics (CER) for evaluation
accelerate>=0.25.0          # Training acceleration utilities
bitsandbytes>=0.41.0        # 8-bit optimizer for memory efficiency
nvtx>=0.2.5                 # NVIDIA Tools Extension for profiling (optional)